# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Analizcim API",
    "description": "Excel analiz uygulaması için REST API. Bu API, satış ve alış dosyalarını analiz etmek, geçmişi yönetmek, tahminler oluşturmak ve kullanıcı yönetimi için kullanılır.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@analizcim.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Yerel geliştirme sunucusu"
    }
  ],
...
  "paths": {
    "/login": {
      "post": {
        "tags": ["Kimlik Doğrulama"],
        "summary": "Kullanıcı girişi",
        "description": "Kullanıcı adı ve şifre ile giriş yapar. Rate limit: 15 dakikada 5 istek.",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Kullanıcı adı",
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "description": "Şifre",
                    "format": "password",
                    "example": "GucluSifre123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Giriş başarılı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "username": {
                          "type": "string",
                          "example": "admin"
                        },
                        "is_admin": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation hatası",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Kullanıcı adı gereklidir."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Geçersiz kimlik bilgileri",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Geçersiz kullanıcı adı veya şifre."
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit aşıldı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Çok fazla giriş denemesi. 15 dakika sonra tekrar deneyin."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": ["Kimlik Doğrulama"],
        "summary": "Yeni kullanıcı kaydı",
        "description": "Yeni bir kullanıcı hesabı oluşturur.",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Kullanıcı adı (3-30 karakter, alfanumerik + alt çizgi)",
                    "example": "yeni_kullanici"
                  },
                  "password": {
                    "type": "string",
                    "description": "Şifre (en az 8 karakter, 1 büyük harf, 1 küçük harf, 1 rakam, 1 özel karakter)",
                    "format": "password",
                    "example": "Sifre123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Kayıt başarılı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 2
                        },
                        "username": {
                          "type": "string",
                          "example": "yeni_kullanici"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation hatası",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Kullanıcı adı 3-30 karakter arasında olmalıdır."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/logout": {
      "post": {
        "tags": ["Kimlik Doğrulama"],
        "summary": "Kullanıcı çıkışı",
        "description": "Mevcut oturumu sonlandırır.",
        "operationId": "logout",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Çıkış başarılı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/change-password": {
      "post": {
        "tags": ["Kimlik Doğrulama"],
        "summary": "Şifre değiştirme",
        "description": "Kullanıcının mevcut şifresini değiştirir.",
        "operationId": "changePassword",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["currentPassword", "newPassword"],
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "description": "Mevcut şifre",
                    "format": "password",
                    "example": "EskiSifre123!"
                  },
                  "newPassword": {
                    "type": "string",
                    "description": "Yeni şifre",
                    "format": "password",
                    "example": "YeniSifre123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Şifre değiştirildi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Şifre başarıyla değiştirildi."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Hatalı şifre veya validation hatası",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Mevcut şifre hatalı."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Oturum kimlik doğrulaması"
      },
      "adminAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Admin yetkisi gerektiren işlemler için"
      }
    }
  }
}